% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_data.R
\docType{data}
\name{example_data}
\alias{example_data}
\title{Example dataset for NullstrapDE}
\format{
A dataset containing 3 objects.
}
\usage{
data(example_data)
}
\description{
This dataset contains:
\describe{
\item{counts}{A simulated count matrix}
\item{colData}{A data frame containing sample-level metadata}
\item{true_de}{Logical vector indicating truly DE genes}
}
}
\keyword{datasets}
